+++ corrscope/gui/__init__.py
@@ -210,6 +210,9 @@ class MainWindow(qw.QMainWindow, Ui_MainWindow):
             self.on_separate_render_dir_toggled
         )
 
+        self.action_parallel.setChecked(self.pref.parallel)
+        self.action_parallel.toggled.connect(self.on_parallel_toggled)
+
         self.action_open_config_dir.triggered.connect(self.on_open_config_dir)
 
         self.actionNew.triggered.connect(self.on_action_new)
@@ -465,6 +468,9 @@ class MainWindow(qw.QMainWindow, Ui_MainWindow):
         else:
             self.pref.render_dir = ""
 
+    def on_parallel_toggled(self, checked: bool):
+        self.pref.parallel = checked
+
     def on_open_config_dir(self):
         appdata_uri = qc.QUrl.fromLocalFile(str(paths.appdata_dir))
         QDesktopServices.openUrl(appdata_uri)
@@ -606,7 +612,10 @@ class MainWindow(qw.QMainWindow, Ui_MainWindow):
             )
 
         arg = Arguments(
-            cfg_dir=self.cfg_dir, outputs=outputs, is_aborted=raise_exception
+            cfg_dir=self.cfg_dir,
+            outputs=outputs,
+            parallelism=self.pref.parallelism(),
+            is_aborted=raise_exception,
         )
         return arg
 
@@ -764,7 +773,7 @@ class CorrThread(Thread):
     job: CorrJob
 
     def __init__(self, cfg: Config, arg: Arguments, mode: PreviewOrRender):
-        Thread.__init__(self)
+        Thread.__init__(self, name="CorrThread")
         self.job = CorrJob(cfg, arg, mode)
 
     def run(self):